Implemented Login in both serif_shared and serif_cli. Greeting is gone, replaced by MatrixClient and MatrixSession. MatrixClient is the API when you're not logged in - getting the version or logging in. Calling login(username, password) on a MatrixClient will return a MatrixSession. MatrixSession has a test() method that sends the test message from our earlier Greeting code.
The serif_cli is set up as a state machine enum called AppState, currently
with two subclasses, Login and Rooms. Login takes a MatrixClient and
handles taking user input to login, then transitions to Rooms with the
returned MatrixSession, which for now simply runs the test() method
on the MatrixSession.

Because when running under Gradle the console object is null,
a run_dist.sh script has also been added that calls gradlew with
:serif_cli:assembleDist and then untars the dist into a test folder, then
runs the app, removing the test folder afterwards if the app exits
sucessfully.

The main work to finish this issue is to add error handling, but this
deserves a discussion. Right now it can throw an unhandled exception.
My intial fix was to use the Kotlin Result<T> type, but it's not currently
allowed as a return type unless you pass an extra flag because it might
change in the future.