Add a context menu with some message actions.
Adds menu options for reply, edit, and reaction, as well as placeholder
entries for delete and show message source. Also enables replies, edits,
and reactions to be sent out. Adds an indicator above the text entry for
showing what type of message is being created for more involved message
types.